chore: use addonUpload for installer and packages
